Addted bot scheduler code
