--- samples/bot-scrum-reminder/csharp/AppCheckinLocation/AdapterWithErrorHandler.cs
@@ -1,56 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using System;
-using System.Threading.Tasks;
-using Microsoft.Bot.Builder;
-using Microsoft.Bot.Connector;
-using Microsoft.Bot.Builder.Integration.AspNet.Core;
-using Microsoft.Bot.Schema;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
-using System.Net.Http;
-
-namespace AppCheckinLocation
-{
-    public class AdapterWithErrorHandler : CloudAdapter
-    {
-        public AdapterWithErrorHandler(IConfiguration configuration, IHttpClientFactory httpClientFactory, ILogger<IBotFrameworkHttpAdapter> logger, ConversationState conversationState = default)
-            : base(configuration, httpClientFactory, logger)
-        {
-            OnTurnError = async (turnContext, exception) =>
-            {
-                // Log any leaked exception from the application.
-                // NOTE: In production environment, you should consider logging this to
-                // Azure Application Insights. Visit https://aka.ms/bottelemetry to see how
-                // to add telemetry capture to your bot.
-                logger.LogError(exception, $"[OnTurnError] unhandled error : {exception.Message}");
-
-                // Send a message to the user
-                await turnContext.SendActivityAsync("The bot encountered an error or bug.");
-                await turnContext.SendActivityAsync("To continue to run this bot, please fix the bot source code.");
-
-                // Send a trace activity, which will be displayed in the Bot Framework Emulator
-                await SendTraceActivityAsync(turnContext, exception);
-            };
-        }
-
-        private static async Task SendTraceActivityAsync(ITurnContext turnContext, Exception exception)
-        {
-            // Only send a trace activity if we're talking to the Bot Framework Emulator
-            if (turnContext.Activity.ChannelId == Channels.Emulator)
-            {
-                Activity traceActivity = new Activity(ActivityTypes.Trace)
-                {
-                    Label = "TurnError",
-                    Name = "OnTurnError Trace",
-                    Value = exception.Message,
-                    ValueType = "https://www.botframework.com/schemas/error",
-                };
-
-                // Send a trace activity
-                await turnContext.SendActivityAsync(traceActivity);
-            }
-        }
-    }
-}
