--- samples/bot-scrum-reminder/csharp/AppCheckinLocation/Bots/ActivityBot.cs
@@ -1,318 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-// Generated with Bot Builder V4 SDK Template for Visual Studio v4.14.0
-
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-using AdaptiveCards;
-using AppCheckinLocation.Models;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Bot.Builder;
-using Microsoft.Bot.Builder.Teams;
-using Microsoft.Bot.Schema;
-using Microsoft.Bot.Schema.Teams;
-using Microsoft.Extensions.Configuration;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
-
-namespace AppCheckinLocation.Bots
-{
-    /// <summary>
-    /// Bot Activity handler class.
-    /// </summary>
-    public class ActivityBot : TeamsActivityHandler
-    {
-        private readonly string _applicationBaseUrl;
-        private readonly IWebHostEnvironment _env;
-        protected readonly BotState _conversationState;
-        protected readonly IStatePropertyAccessor<UserLocationDetail> _UserDetail;
-
-        public ActivityBot(IConfiguration configuration, IWebHostEnvironment env, ConversationState conversationState)
-        {
-            _conversationState = conversationState;
-            _env = env;
-            _applicationBaseUrl = configuration["ApplicationBaseUrl"] ?? throw new NullReferenceException("ApplicationBaseUrl");
-            _UserDetail = conversationState.CreateProperty<UserLocationDetail>(nameof(UserLocationDetail));
-        }
-
-        /// <summary>
-        /// Handle when a message is addressed to the bot.
-        /// </summary>
-        /// <param name="turnContext">The turn context.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        /// <returns>A task that represents the work queued to execute.</returns>
-        protected override async Task OnMessageActivityAsync(ITurnContext<IMessageActivity> turnContext, CancellationToken cancellationToken)
-        {
-            if(turnContext.Activity.Text.ToLower().Trim() == "create-reminder")
-            {
-                await turnContext.SendActivityAsync(MessageFactory.Attachment(GetAdaptiveCardForTaskModule()), cancellationToken);
-            }           
-        }
-
-        /// <summary>
-        /// Handle request from bot.
-        /// </summary>
-        /// <param name = "turnContext" > The turn context.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        /// <returns>A task that represents the work queued to execute.</returns>
-        public override async Task OnTurnAsync(ITurnContext turnContext, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            await base.OnTurnAsync(turnContext, cancellationToken);
-
-            // Save any state changes that might have occurred during the turn.
-            await _conversationState.SaveChangesAsync(turnContext, false, cancellationToken);
-        }
-
-        /// <summary>
-        /// Invoked when bot (like a user) are added to the conversation.
-        /// </summary>
-        /// <param name="membersAdded">A list of all the members added to the conversation.</param>
-        /// <param name="turnContext">Context object containing information cached for a single turn of conversation with a user.</param>
-        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>A task that represents the work queued to execute.</returns>
-        /// <remarks>
-        protected override async Task OnMembersAddedAsync(IList<ChannelAccount> membersAdded, ITurnContext<IConversationUpdateActivity> turnContext, CancellationToken cancellationToken)
-        {
-            foreach (var member in turnContext.Activity.MembersAdded)
-            {
-                if (member.Id != turnContext.Activity.Recipient.Id)
-                {
-                    await turnContext.SendActivityAsync(MessageFactory.Text($"Hello and welcome! With this sample you can schedule a task and get reminder on the scheduled date and time.(use command 'create-reminder')."), cancellationToken);
-                }
-            }
-        }
-
-        /// <summary>
-        /// Handle task module is fetch.
-        /// </summary>
-        /// <param name = "turnContext" > The turn context.</param>
-        /// <param name = "taskModuleRequest" >The task module invoke request value payload.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        /// <returns>A Task Module Response for the request.</returns>
-        protected override Task<TaskModuleResponse> OnTeamsTaskModuleFetchAsync(ITurnContext<IInvokeActivity> turnContext, TaskModuleRequest taskModuleRequest, CancellationToken cancellationToken)
-        {
-            var asJobject = JObject.FromObject(taskModuleRequest.Data);
-            var buttonType = (string)asJobject.ToObject<CardTaskFetchValue<string>>()?.Id;
-            var taskModuleResponse = new TaskModuleResponse();
-
-            if (buttonType == "schedule")
-            {
-                taskModuleResponse.Task = new TaskModuleContinueResponse
-                {
-                    Type = "continue",
-                    Value = new TaskModuleTaskInfo()
-                    {
-                        Url = _applicationBaseUrl + "/" + "CheckIn",
-                        Height = 350,
-                        Width = 350,
-                        Title = "Schedule Task",
-                    },
-                };
-            }
-
-            return Task.FromResult(taskModuleResponse);
-        }
-
-        /// <summary>
-        /// Handle task module is submit.
-        /// </summary>
-        /// <param name = "turnContext" > The turn context.</param>
-        /// <param name = "taskModuleRequest" >The task module invoke request value payload.</param>
-        /// <param name="cancellationToken">The cancellation token.</param>
-        /// <returns>A Task Module Response for the request.</returns>
-        protected override async Task<TaskModuleResponse> OnTeamsTaskModuleSubmitAsync(ITurnContext<IInvokeActivity> turnContext, TaskModuleRequest taskModuleRequest, CancellationToken cancellationToken)
-        {
-            var locationInfo = JObject.FromObject(taskModuleRequest.Data);
-
-            Console.WriteLine("Hello task");
-
-            await turnContext.SendActivityAsync(MessageFactory.Attachment(GetAdaptiveCardForTaskModule()), cancellationToken);
-
-            Scheduler sc = new Scheduler();
-            sc.Start();
-
-            return null;
-        }
-
-        /// <summary>
-        /// Sample Adaptive card for check in button.
-        /// </summary>
-        private Attachment GetAdaptiveCardForTaskModule()
-        {
-            AdaptiveCard card = new AdaptiveCard(new AdaptiveSchemaVersion("1.2"))
-            {
-                Body = new List<AdaptiveElement>
-                {
-                    new AdaptiveTextBlock
-                    {
-                        Text = "Please click on schedule to schedule task",
-                        Weight = AdaptiveTextWeight.Bolder,
-                        Spacing = AdaptiveSpacing.Medium,
-                    }
-                },
-                Actions = new List<AdaptiveAction>
-                {
-                    new AdaptiveSubmitAction
-                    {
-                        Title = "Schedule task",
-                        Data = new AdaptiveCardAction
-                        {
-                            MsteamsCardAction = new CardAction
-                            {
-                                Type = "task/fetch",
-                            },
-                            Id="schedule"
-                        },
-                    }
-                },
-            };
-
-            return new Attachment()
-            {
-                ContentType = AdaptiveCard.ContentType,
-                Content = card,
-            };
-        }
-
-        ///// <summary>
-        ///// Sample Adaptive card for user current location info.
-        ///// </summary>
-        //private Attachment GetAdaptiveCardForUserLocation(string time, string user, double latitude, double longitude)
-        //{
-        //    AdaptiveCard card = new AdaptiveCard(new AdaptiveSchemaVersion("1.2"))
-        //    {
-        //        Body = new List<AdaptiveElement>
-        //        {
-        //            new AdaptiveTextBlock
-        //            {
-        //                Text = $"User name :{user}",
-        //                Weight = AdaptiveTextWeight.Bolder,
-        //                Spacing = AdaptiveSpacing.Medium,
-        //                Wrap = true
-        //            },
-        //            new AdaptiveTextBlock
-        //            {
-        //                Text = $"Check in time: {time}",
-        //                Weight = AdaptiveTextWeight.Bolder,
-        //                Spacing = AdaptiveSpacing.Medium,
-        //                Wrap = true
-        //            }
-        //        },
-        //        Actions = new List<AdaptiveAction>
-        //        {
-        //            new AdaptiveSubmitAction
-        //            {
-        //                Title = "View location",
-        //                Data = new AdaptiveCardAction
-        //                {
-        //                    MsteamsCardAction = new CardAction
-        //                    {
-        //                        Type = "task/fetch",
-        //                    },
-        //                    Id = "viewLocation",
-        //                    Latitude = latitude,
-        //                    Longitude = longitude
-        //                },
-        //            }
-        //        },
-        //    };
-
-        //    return new Attachment()
-        //    {
-        //        ContentType = AdaptiveCard.ContentType,
-        //        Content = card,
-        //    };
-        //}
-
-        /// <summary>
-        /// Sample Adaptive card for user's last check in's.
-        /// </summary>
-        //private List<Attachment> GetAdaptiveCardForUserLastCheckIn(List<UserDetail> userDetails)
-        //{
-        //    List<Attachment> attachmentList = new List<Attachment>();
-
-        //    foreach (var user in userDetails)
-        //    {
-        //        AdaptiveCard card = new AdaptiveCard(new AdaptiveSchemaVersion("1.2"))
-        //        {
-        //            Body = new List<AdaptiveElement>
-        //            {
-        //                new AdaptiveTextBlock
-        //                {
-        //                    Text = $"User name :{user.UserName}",
-        //                    Weight = AdaptiveTextWeight.Bolder,
-        //                    Spacing = AdaptiveSpacing.Medium,
-        //                    Wrap = true
-        //                },
-        //                new AdaptiveTextBlock
-        //                {
-        //                    Text = $"Check in time: {user.CheckInTime}",
-        //                    Weight = AdaptiveTextWeight.Bolder,
-        //                    Spacing = AdaptiveSpacing.Medium,
-        //                    Wrap = true
-        //                }
-        //            },
-        //            Actions = new List<AdaptiveAction>
-        //            {
-        //                new AdaptiveSubmitAction
-        //                {
-        //                    Title = "View location",
-        //                    Data = new AdaptiveCardAction
-        //                    {
-        //                        MsteamsCardAction = new CardAction
-        //                        {
-        //                            Type = "task/fetch",
-        //                        },
-        //                        Id = "viewLocation",
-        //                        Latitude = user.Latitude,
-        //                        Longitude = user.Longitude
-        //                    },
-        //                }
-        //            },
-        //        };
-
-        //        Attachment attachment = new Attachment()
-        //        {
-        //            ContentType = AdaptiveCard.ContentType,
-        //            Content = card,
-        //        };
-
-        //        attachmentList.Add(attachment);
-        //    }
-
-        //    return attachmentList;
-        //}
-
-        // Save user details in json file.
-        private async Task SaveUserDetailsAsync(ITurnContext<IInvokeActivity> turnContext, UserDetail userDetails)
-        {
-            var currentUserDetail = await this._UserDetail.GetAsync(turnContext, () => new UserLocationDetail());
-            List<UserDetail> userDetailsList = new List<UserDetail>();
-            if (currentUserDetail.UserDetails == null)
-            {
-                userDetailsList.Add(userDetails);
-                currentUserDetail.UserDetails = userDetailsList;
-                await this._UserDetail.SetAsync(turnContext, currentUserDetail);
-            }
-            else if (currentUserDetail.UserDetails.Count == 10)
-            {
-                currentUserDetail.UserDetails.RemoveAt(0); 
-                userDetailsList = currentUserDetail.UserDetails;
-                userDetailsList.Add(userDetails);
-                currentUserDetail.UserDetails = userDetailsList;
-                await this._UserDetail.SetAsync(turnContext, currentUserDetail);
-            }
-            else
-            {
-                userDetailsList = currentUserDetail.UserDetails;
-                userDetailsList.Add(userDetails);
-                currentUserDetail.UserDetails = userDetailsList;
-                await this._UserDetail.SetAsync(turnContext, currentUserDetail);
-            }
-        }
-    }
-}
