--- samples/bot-scrum-reminder/csharp/AppCheckinLocation/Job.cs
@@ -1,16 +1,0 @@
-﻿using Quartz;
-using System;
-using System.Threading.Tasks;
-
-namespace AppCheckinLocation
-{
-    public class Job: IJob
-    {
-        public Task Execute(IJobExecutionContext context)
-        {
-            Console.WriteLine("Hello Job");
-
-            return null;
-        }
-    }
-}
