--- samples/bot-scrum-reminder/csharp/AppCheckinLocation/Models/CardTaskFetchValue.cs
@@ -1,23 +1,0 @@
-﻿// <copyright file="CardTaskFetchValue.cs" company="Microsoft">
-// Copyright (c) Microsoft. All rights reserved.
-// </copyright>
-
-using Newtonsoft.Json;
-
-namespace AppCheckinLocation.Models
-{
-    /// <summary>
-    /// Card task fetch value model class.
-    /// </summary>
-    public class CardTaskFetchValue<T>
-    {
-        [JsonProperty("type")]
-        public object Type { get; set; } = "task/fetch";
-
-        [JsonProperty("id")]
-        public object Id { get; set; }
-
-        [JsonProperty("data")]
-        public T Data { get; set; }
-    }
-}
