--- samples/bot-scrum-reminder/csharp/AppCheckinLocation/Pages/CheckIn.cshtml.cs
@@ -1,12 +1,0 @@
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace AppCheckinLocation.Pages
-{
-    public class CheckInModel
-        : PageModel
-    {
-        public void OnGet()
-        {
-        }
-    }
-}
