--- samples/bot-scrum-reminder/csharp/AppCheckinLocation/Pages/ViewLocation.cshtml
@@ -1,57 +1,0 @@
-﻿@page
-@model AppCheckinLocation.Pages.ViewLocationModel
-@{
-}
-<head>
-    <script src="https://statics.teams.cdn.office.net/sdk/v1.10.0/js/MicrosoftTeams.min.js"
-            integrity="sha384-6oUzHUqESdbT3hNPDDZUa/OunUj5SoxuMXNek1Dwe6AmChzqc6EJhjVrJ93DY/Bv"
-            crossorigin="anonymous"></script>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
-    <script>
-        microsoftTeams.initialize(()=>{
-            viewLocation(); 
-        })
-
-        // Method to show geo location for given latitude and longitude values.
-        function viewLocation() {
-            const params = window.location.search;
-            let latitude = params.substring(params.search("=")+1,params.search("&"));
-            let longitude = params.substring(params.search("&")+11);
-            let location = {"latitude":latitude,"longitude":longitude};
-
-            // Methos to ask for permission and then show current user location
-            microsoftTeams.location.showLocation(location, (error, result) => {
-
-                // If there's any error, an alert shows the error message/code
-                if (error)
-                {
-                    if (error.message)
-                    {
-                        alert(" ErrorCode: " + error.errorCode + error.message);
-                    }
-                    else 
-                    {
-                        alert(" ErrorCode: " + error.errorCode);
-                    }
-
-                    return;
-                }
-            });
-        } 
-    </script>
-    <style>
-        .submitClick{
-            margin:5rem;
-            margin-left:0rem;
-        }
-    </style>
-</head>
-<body class="theme-light">
-    <div class="surface">
-        <div class="panel">
-            <div>
-            </div>
-        </div>
-    </div>
-</body>
