--- samples/bot-scrum-reminder/csharp/AppCheckinLocation/Pages/ViewLocation.cshtml.cs
@@ -1,12 +1,0 @@
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace AppCheckinLocation.Pages
-{
-    public class ViewLocationModel
-        : PageModel
-    {
-        public void OnGet()
-        {
-        }
-    }
-}
