--- samples/bot-scrum-reminder/csharp/AppCheckinLocation/appsettings.json
@@ -1,5 +1,0 @@
-{
-  "MicrosoftAppId": "68af893d-75e2-4277-857e-3105101e9a8b",
-  "MicrosoftAppPassword": "KBOa_8c7O~BXbv1c.s07YNZTtJ.TPXlF.2",
-  "ApplicationBaseUrl": "https://d37c-116-75-27-174.ngrok.io"
-}
