+++ samples/bot-task-reminder/csharp/BotTaskReminder/AppPackage/color.png
