+++ samples/bot-task-reminder/csharp/BotTaskReminder/AppPackage/outline.png
