+++ samples/bot-task-reminder/csharp/BotTaskReminder/BotTaskReminder.csproj
