+++ samples/bot-task-reminder/csharp/BotTaskReminder/Pages/ScheduleTask.cshtml.cs
@@ -1,0 +1,12 @@
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace AppCheckinLocation.Pages
+{
+    public class ScheduleTaskModel
+        : PageModel
+    {
+        public void OnGet()
+        {
+        }
+    }
+}
