+++ samples/bot-task-reminder/csharp/BotTaskReminder/Properties/launchSettings.json
