+++ samples/bot-task-reminder/csharp/BotTaskReminder/appsettings.Development.json
