+++ samples/bot-task-reminder/csharp/BotTaskReminder/appsettings.json
@@ -1,0 +1,5 @@
+{
+  "MicrosoftAppId": "{{Microsoft-App-Id}}",
+  "MicrosoftAppPassword": "{{ Microsoft-App-Password}}",
+  "ApplicationBaseUrl": "{{ Application Base Url }}"
+}
